wip: impact calculations, refactoring

+++ domain/impact.go
@@ -3,7 +3,7 @@ package domain
 import "fmt"
 
 type Impact struct {
-	Duration float32 `json:"duration,omitempty"`
+	Duration int     `json:"duration,omitempty"`
 	Chance   float32 `json:"chance"`
 }
 
@@ -14,9 +14,9 @@ type DamageImpact struct {
 
 func (d DamageImpact) String() string {
 	if d.Chance != 0 {
-		return fmt.Sprintf("{%s, chance: %g, duration: %g}", d.Damage.String(), d.Impact.Chance, d.Impact.Duration)
+		return fmt.Sprintf("{%s, chance: %g, duration: %d}", d.Damage.String(), d.Impact.Chance, d.Impact.Duration)
 	} else {
-		return fmt.Sprintf("{%s, duration: %g}", d.Damage.String(), d.Impact.Duration)
+		return fmt.Sprintf("{%s, duration: %d}", d.Damage.String(), d.Impact.Duration)
 	}
 }
